Last fibonacci

+++ 0x02-functions_nested_loops/104-fibonacci.c
@@ -12,7 +12,7 @@ int main(void)
 {
 	int i;
 
-	unsigned int a, b, c;
+	unsigned long int a, b, c;
 
 	a = 1;
 	b = 2;
